test(a2a): fold streaming coverage into the mapped a2a chat test file
CLAUDE.md asks bug fixes to extend the existing mapped test file for the
directory rather than add a new module. Moves the streaming-iterator
regression tests into tests/test_litellm/llms/a2a/chat/test_a2a_chat_transformation.py
so the a2a chat bridge keeps one owner, and widens that file's docstring to
match.

Reported by Greptile on #39513.
